Publish on specified Exchange.
publish() can optionally take an exchange name. But if not provided,
it will publish on the default "nameless" exchange. The semantics
for the nameless exchange are different for routing:

"The exchange parameter is the the name of the exchange.
The empty string denotes the default or nameless exchange:
messages are routed to the queue with the name
specified by routing_key, if it exists."